raspberry_scripts/requirements.txt
pyserial>=3.5
redis>=7.2
opencv-python>=4.13
paho-mqtt>=2.1